requirements.txt
torch
transformers
pandas
numpy
scikit-learn
jupyterlab
python-dotenv
sentencepiece
protobuf